image-view, tool-crop: Use double instead of float in the public API
Even though we don't need the extra accuracy of double-precision
floating-point numbers, it doesn't cost us that much either because we
are not going to have hundreds of PhotosImageView instances. On the
other hand it does give us consistency with the Cairo and GVariant
APIs, and the future GtkImageView API that's in the works.

+++ src/photos-image-view.h
@@ -35,11 +35,11 @@ GtkWidget          *photos_image_view_new_from_node      (GeglNode *node);
 
 GeglNode           *photos_image_view_get_node           (PhotosImageView *self);
 
-gfloat              photos_image_view_get_x              (PhotosImageView *self);
+gdouble             photos_image_view_get_x              (PhotosImageView *self);
 
-gfloat              photos_image_view_get_y              (PhotosImageView *self);
+gdouble             photos_image_view_get_y              (PhotosImageView *self);
 
-gfloat              photos_image_view_get_zoom           (PhotosImageView *self);
+gdouble             photos_image_view_get_zoom           (PhotosImageView *self);
 
 void                photos_image_view_set_node           (PhotosImageView *self, GeglNode *node);
